requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
apscheduler==3.10.4
sqlalchemy==2.0.23
alembic==1.13.1
